Problem 30: Digit n powers
Surprisingly there are only three numbers that can be written as the sum of fourth powers of their digits:

1634 = 14 + 64 + 34 + 44
8208 = 84 + 24 + 04 + 84
9474 = 94 + 44 + 74 + 44
As 1 = 14 is not a sum it is not included.

The sum of these numbers is 1634 + 8208 + 9474 = 19316.

Find the sum of all the numbers that can be written as the sum of n powers of their digits.

Tests: 

Test 1.  digitnPowers(2) should return a number.
Test 2.  digitnPowers(2) should return 0.
Test 3.  digitnPowers(3) should return 1301.
Test 4  digitnPowers(4) should return 19316.
Test 5.  digitnPowers(5) should return 443839.