Create Problem024: Lexicographic permutations
A permutation is an ordered arrangement of objects. For example, 3124 is one possible permutation of the digits 1, 2, 3 and 4. If all of the permutations are listed numerically or alphabetically, we call it lexicographic order. The lexicographic permutations of 0, 1 and 2 are:

012   021   102   120   201   210
What is the nth lexicographic permutation of the digits 0, 1, 2, 3, 4, 5, 6, 7, 8 and 9?

Tests

Test 1. lexicographicPermutations(699999) should return a number.
Test 2. lexicographicPermutations(699999) should return 1938246570.
Test 3. lexicographicPermutations(899999) should return 2536987410.
Test 4. lexicographicPermutations(900000) should return 2537014689.
Test 5. lexicographicPermutations(999999) should return 2783915460.